+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.9.1
+- Add docs.rs metadata
+
+## 0.9.0
+- Remove `image` default features
+- Put `print_from_file` behind the new `print-file` feature flag
+
+## 0.8.1
+- Revert removed `image` features
+
+## 0.8.0
+- Remove unneeded features and update dependencies
+- Use Catmull-Rom for up/downscaling
+- Add `premultiplied_alpha` Config option
+
 ## 0.7.1
 - Bump `base64` and `crossterm` dependencies